src/util/background-gl.cpp:  implement "tiled" with gnome-bg functions
when 1) compiled with -DGNOME_BG, and 2) option background.fill_mode
is set to "preserve_aspect" the GNOME_BG codepath is activated.

Options in the GNOME_BG codepath:

If "tile" is "true", the image is tiled to fill the screen.

If "span" is "false" the image is read by the gdkpixbuf library in its
natural dimensions instead of being limited to the window
dimensions.

If "always_fit" is "true" and "span" is "false", the image will be
always be downscaled to fit within the screen's window dimensions.

The "always_fit" option makes sense only when "span" is false.

("preserve_aspect" is not a boolean option anymore, the aspect is always
preserved.)