src/background/background.cpp: (load_next_background): don't destroy window on failure
Our use of GFileMonitor is very inefficient (it adds a new monitor on
every change to the target image file, and rescans the directories on
every change) - but the advantage is that you can change the
target image file to a directory and vice versa, and add or remove
files from the directory and wf-background should automatically pick
up those changes.

For this to work the GtkWindow must not be removed on some error. To
mitigate the downside of a "stale" background image that may be displayed
(after an error encoutered in displaying an image),
create a pixmap in WayfireBackground::set_background and blank it

[250722] - upstream fixed window removal in unrelated gtk4 commit
1ea0ce.  the other code is not relevant:

```
    if (!pbuf)
    {
        pbuf = Gdk::Pixbuf::create(Gdk::Colorspace::COLORSPACE_RGB,
            true,
            8,
            window.get_allocated_width() * scale,
            window.get_allocated_height() * scale);
        pbuf->fill(0x00000000);
    }
```