src/background/background.cpp: (load_next_background): don't destroy window on failure
Our use of GFileMonitor is very inefficient (it adds a new monitor on
every change to the target image file, and rescans the directories on
every change) - but the advantage is that you can change the
target image file to a directory and vice versa, and add or remove
files from the directory and wf-background should automatically pick
up those changes.

For this to work the GtkWindow must not be removed on some error. To
mitigate the downside of a "stale" background image that may be displayed
(after an error encoutered in displaying an image),
create a pixmap in WayfireBackground::set_background and blank it