background: new options when using GNOME_BG
Add three new boolean-valued options under "background" - "tile",
"center" and "span". [230225] implement it. [230226] implement
"always_fit".

If "span" is "false" the image is read by the gdkpixbuf library in
its natural dimensions instead of being limited to the window
dimensions. If "tile" is "true", images are tiled to fill the screen.
If "center" is "true" the image is centered on the screen.

The "always_fit" option makes sense only when "span" is false.
if "always_fit" is "true" and "span" is "false", the image will be
always be downscaled to fit within the screen's window dimensions.

It should be possible to edit background.{tile,center,span} in
wf-shell.ini and see the changes happening on the next
cycle_timeout. (or on changing the disk target image via
file-monitored changes)

* metadata/background.xml.in: add new options. (unfortunately can't
define these conditionally on #ifdef GNOME_BG)

* src/background.hpp: add new options.

* src/background.cpp: if GNOME_BG is defined, do not follow the older
code path.  (pixbuf_tile): now grows two new hairy boolean arguments
`centered', and `notiled'. (WayfireBackground::create_from_file_safe):
use it here.

;240514  use wayfire options in conjunction with GNOME_BG

;240507 - replicate code out of ifdefs to pacify uncrusify