background: new options when using GNOME_BG
Add three new boolean-valued options under "background" - "tile",
"centerand "span".

If "span" is "false" the image is read by the gdkpixbuf library in
its natural dimensions instead of being limited to the window
dimensions. If "tile" is "true", images are tiled to fill the screen.
If "center" is "true" the image is centered on the screen. (TODO)

* metadata/background.xml.in: add new options. (unfortunately can't define
these conditionally on #ifdef GNOME_BG)

* src/background.hpp: add new options.

* src/background.cpp: if GNOME_BG is defined, do not follow the older
code path